+++ src/cpp/session/modules/SessionUserPrefValues.R
@@ -2460,6 +2460,16 @@
    clear = function() { .rs.clearUserPref("code_formatter_external_command") }
 )
 
+# Use Air for code formatting
+#
+# When set, RStudio will automatically select and use an appropriate version of
+# Air when formatting code in projects containing an air.toml file.
+.rs.uiPrefs$useAirFormatter <- list(
+   get = function() { .rs.getUserPref("use_air_formatter") },
+   set = function(value) { .rs.setUserPref("use_air_formatter", value) },
+   clear = function() { .rs.clearUserPref("use_air_formatter") }
+)
+
 # Reformat documents on save
 #
 # When set, the selected formatter will be used to reformat documents on save.